api/requirements.txt
Django==2.1.5
django-filter==2.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.7
pytz==2018.9
